+++ lib/stamina/utils.rb
@@ -1,0 +1,1 @@
+require 'stamina/utils/decorate'